+++ src/components/loginForm.jsx
@@ -1,9 +1,9 @@
-import React, { Component } from "react";
+import React from "react";
 import Joi from "joi-browser";
-import Input from "./common/input";
+import Form from "./common/form";
 
-class LoginForm extends Component {
-  state = { account: { username: "", password: "" }, errors: {} };
+class LoginForm extends Form {
+  state = { data: { username: "", password: "" }, errors: {} };
 
   schema = {
     username: Joi.string()
@@ -14,49 +14,8 @@ class LoginForm extends Component {
       .label("Password")
   };
 
-  validate = () => {
-    const { error } = Joi.validate(this.state.account, this.schema, {
-      abortEarly: false
-    });
-
-    if (!error) return null;
-
-    const errors = {};
-
-    for (let item of error.details) {
-      errors[item.path[0]] = item.message;
-    }
-
-    return errors;
-  };
-
-  validateProperty = ({ name, value }) => {
-    const obj = { [name]: value };
-    const schema = { [name]: this.schema[name] };
-    const { error } = Joi.validate(obj, schema);
-    return error ? error.details[0].message : null;
-  };
-
-  handleSubmit = e => {
-    e.preventDefault();
-
-    const errors = this.validate();
-    this.setState({ errors: errors || {} });
-    if (errors) return;
-  };
-
-  handleChange = ({ target: input }) => {
-    const errors = { ...this.state.errors };
-    const errorMessage = this.validateProperty(input);
-    if (errorMessage) errors[input.name] = errorMessage;
-    else delete errors[input.name];
-
-    const account = { ...this.state.account };
-    account[input.name] = input.value;
-    this.setState({
-      account,
-      errors
-    });
+  doSubmit = () => {
+    console.log("submited");
   };
 
   render() {
@@ -64,29 +23,9 @@ class LoginForm extends Component {
       <div>
         <h2>Login</h2>
         <form onSubmit={this.handleSubmit}>
-          <Input
-            name="username"
-            label="Username"
-            value={this.state.account.username}
-            error={this.state.errors.username}
-            onChange={this.handleChange}
-            focus={true}
-          />
-          <Input
-            name="password"
-            label="Password"
-            type="password"
-            value={this.state.account.password}
-            error={this.state.errors.password}
-            onChange={this.handleChange}
-          />
-          <button
-            disabled={this.validate()}
-            type="submit"
-            className="btn btn-primary"
-          >
-            Submit
-          </button>
+          {this.renderInput("username", "Username", undefined, true)}
+          {this.renderInput("password", "Password", "password")}
+          {this.renderButton("Login")}
         </form>
       </div>
     );
